requirements.txt
python-telegram-bot[job-queue, webhooks]>=20.8
openai>=1.3
gspread>=5.12
oauth2client>=4.1
Flask>=3.0
functions-framework>=3.4
requests>=2.31
